Backend/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.1
